LayoutTests/http/tests/security/cross-frame-access-frameelement-expected.txt

CONSOLE MESSAGE: Unsafe JavaScript attempt to access frame with URL http://127.0.0.1:8000/security/cross-frame-access-frameelement.html from frame with URL http://127.0.0.1:8080/security/resources/cross-frame-access-frameelement-from-iframe.html. Domains, protocols and ports must match.

This test checks if frameElement is accessible from the same or cross origin iframe.

 

--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
[object HTMLIFrameElement]

--------
Frame: '<!--framePath //<!--frame1-->-->'
--------
undefined
